rework npm packages
npm and clojars should now always be release at the same time

the launcher jar is independent and will probably be updated much
less frequently. need to figure out if there is a way to get a complete
dependency graph for a clojars release via some sort of web api.

then we wouldn't need the launcher jar at all