shadow-cljs 2.0 - the refactoring
This work started as a simple branch trying to make it easier to
work with JS dependencies either via Closure or other mechanisms.

Instead it ended up in an entire rework of most internals to
streamline that work and making the entire system more flexible.

The work is not finished yet but is in a usable state. Since I
messed up some branch merges I'm not doing this as a PR.

The public API remains the same and the internals have changed.