don't use a goog define to track jar version in CLI script
getting rather annoying to recompile the CLI script to
release updates if all I need to change is the version.

It now uses the package.json via jar-version property. It does not
use "version" since I might need to release an update to the CLI
script without updating the shadow-cljs jar.