remove global protection for node targets
CLJS or the Closure Library both don't try to protect the global
so neither do we. It is only an issue when running two dev builds
in the same VM which should never be done anyways.

It was created due to conflicts when build atom plugins in CLJS
while using ProtoREPL. We can now use :npm-module to do this
safely.

closes #137