fix watch* to use passed-in build-config
Currently if you pass in a config map to `watch*`, it is ignored except for its `:build-id` key which is used to read the config again. This patch would continue to allow `watch*` to accept _either_ a build-id or build-config map. An alternative would be to remove this part altogether so that `watch*` _requires_ a config map.