create an AOT compiled artifact for the standalone version
startup is way faster with AOT compiled code and should be ok
for most CLJS only projects. the default jar does not contain
any AOT compiled code and should be safer to use when
embedding in lein & co.

The npm version will use the AOT artifact by default and this
can be disabled by setting `:aot false` in the config.