fix node-repl eval issue
code that generates a leading docstring breaks the constructed
JS to capture the result value.

  (defn x "foo" [a] a)

works fine in the browser but in node an extra wrapper fn is
added after the fact that gets confused by the multiline comment

  /*
   * foo
   */

that is added in front of the generated JS. quick fix just strips
and comment generated in front.