begin changing REPL init logic
first part is not making the REPL code part of the build
but instead have the REPL initialize itself when the runtime
first connects.

Can't add cljs.user to the base module by default since that
messes up :module split logic if the REPL uses it to load things.

vim-fireplace does (ns cljs.user (:require a-user-ns)) whenever
it loads a namespace. this causes cljs.user to depend on that
ns and causing problems if that ns is not in the base module.

Moving the REPL logic closer to the actual REPL and out of the
target responsibilty this should be avoided.

fixes #667