fix incremental JS compile issue
can't really do incremental compiles anymore. closure-compiler
is too picky about renaming rules and I can't find a way to cache
exports info for all namespaces or how to restore it so the
compiler actually uses it. fails with weird internal errors
when it doesn't get all the sources. the previous hacks used for
this no longer work.

only affects classpath JS so shouldn't be too much of an issue