[BREAKING] fix import of npm package in classpath-js
it used to work in dev builds but broke in release because of
an inconsistent use of import patterns. now only supports

   import Thing from "npm-package";

the previous

   import * as Thing from "npm-package";

will now fail. Unfortunately there was no way to make this work
otherwise because of the strict checking the Closure Compiler
does. It does however now work in release and dev.