Strip 'Secure' flag from cookies passing through dev-http proxy
Developers using the dev-proxy with http against an https backend may
have trouble if the backend submits cookies with the 'Secure' flag set.
Certain browsers, such as Chrome, will balk at cookies with 'Secure' set
that arrive via http.  This patch enables this type of flow to proceed
by removing the Secure token as it passes through the dev-proxy.

Signed-off-by: Greg Haskins <greg@manetu.com>