remove hawk library used for file watching
Only used to be used on macOS since it was slightly faster
than the default JVM implementation. However in Big Sur it seems
to cause issues and break completely or just be a lot slower.

Win/Linux are unaffected since they always used the default JVM
implemenation. Big Sur seems to be faster. Previous macOS version
might get a tad slower but I haven't tested that in a long time
and it used to be no more than 2secs which is fine. Better than
relying on an unmaintained possibly broken library.

fixes #837
fixes #774