fix ESM .js code referencing other ESM .js code
Closure assumes that the final output is all in the same scope
but for node that isn't true currently. So instead of trying to
change the way Closure rewrites we just collect the names it
creates and lift those into the global scope when needed.

only really affected watch/compile builds. release doesn't have
this problem.

fixes #836
fixes #521