fix inspect blowing up on misbehaving types
datascript/DB pretends to be a record but (keys db) breaks.
Inspect runs into this since (map? db) is true and it tries to
sort the keys for display purposes which fails with in interal
exception since they (seq db) isn't a seq of MapEntry instances.

Don't know if this should be a considered a bug in Datascript but
it only surfaced because inspect calls (keys db) which is user
is unlikely to do. Since this throws the inspect just adds a
try/catch (which it should have had anyways) and treat the type
as unsupported instead.

fixes #858