bump default :output-feature-set to :es6
really doesn't make sense to have :es5 as a default still. ES6 is
from 2015 and widely supported for years. Only exception was
IE but that is dead. This likely is still too conservative and
it might even be fine to bump to :es8 default. Users that need
support for older versions can always configure it.