treat symbols as valid requires when listed in npm-deps
:js-provider :require needs to know which symbols are valid npm
packages, eg. (:require [react]). For this it currently has to
look at node_modules which is not ideal if the package is
actually installed in some nested folder.

It'll now gather :npm-deps listed in deps.cljs files from the
classpath and treat those as valid symbols without requiring
them to be actually present on disk.