enhance externs inference for aliased types
(def Thing js/Date) or (def Thing some-npm/Thing)

will have some.ns/Thing as :tag in places such as extend-protocol.
Since this is a full tag but not a JS tag it'll be ignored by
externs inference assuming it can be renamed. I think it might be
a safer default to have it not rename stuff instead. Unlikely
to do a lot of interop on JS types with properties that can be
renamed.

fixes #897