change default npm resolve rules to match node/webpack
shadow-cljs previously did not support nested package installs.
which on one hand means that it won't include packages twice but
on the other could run into unresolvable dependency conflicts.

given that this is the default in all JS tools and required for
some projects that have unresolvable dependency conflicts and
must include a library twice to work.

this can be opted out of by setting

   :js-options {:allow-nested-packages false}

I opted to make this the default in shadow-cljs too since it is
more likely to "fix" builds than it is to break them. It may
however now lead to builds using multiple versions of certain
packages when it previously wouldn't. I will add warnings for this
case in a future update.

fixes #892