fix greedy builds not actually picking up new files
greedy builds are those that pick up new files by name without
any direct require (eg. :ns-regexp). new files may match so they
need to be aware of new files.

on some OS the :new event is immediately followed by :mod. given
that those notifications are stored in a map the :mod overwrites
the :new and the info that this is actually a new file is lost.