dependency trim: pathom, remove graph api
the current graph api was only used to initially fetch some data
for the UI. this is now just a regular request. using pathom for
this was absolute overkill and most of the other graph stuff was
broken and unused anyways.

pathom adds quite a bit of startup time with currently no use. so
it is removed without replacement. it'll come back in some other
form later maybe when I actually have a proper use for it.