actually still against polluting cljs.reader by default
setting

  :compiler-options {:data-readers true}

will lead to cljs.reader/*tag-table* being populated with the
readers listed in data_readers.clj(c) on the classpath during
compilation.

Not doing this by default since there is no way to tell
if those readers are actually required. Just putting an unresolved
symbol reference in the build makes closure unhappy if the
namespace is not part of the actual build otherwise.

cljs does this by default. I don't like breaking that behavior
but I really don't like this default. Adding a data reader
should be part of user code, not due to some random artifact
being on the classpath.

Does not affect how data readers are handled for compilation or
REPL purposes.