add config option to disable namespace reset in watch
`watch` build normally delete all analyzer data for a given
namespace when getting recompiled and hot-loaded. This is done
so you immediately get warnings for lingering references to
renamed/deleted vars.

Sometimes this can interfere with REPL workflows that want
stuff to stay as is. The runtime actually never deletes references
anyways, so just keeping the analyzer data around when enabled
is enough.

fixes #1025