don't warn on hidden protocol lookups
the compiler uses tags when emitting protocol invocations to
optionally emit a direct call when the targed is known to
implement a protocol. this can lead to a warning in a hot-reload
context where a tag might be undefined at the time of compilation

should not affect anything other than those hot-reload
situations, so we can safely turn of the warning as the emitted
code is correct either way.

fixes #1149