relax :npm-module to remove some hackiness
it is no longer a goal trying to isolate everything and avoiding
the global scope. I doubt anyone ever really cared and it made
everything so damn unreliable in the REPL/hot-reload.

release builds are still isolated and not global, dev builds
just use the global scope as they would in other targets.