add experimental constant lifting
things like

   (get-in thing [:foo :bar])

will allocate the vector each time this is called. the change
allows the compiler to extract this so its only allocated once,
which in theory should speed things up. clojure also does this.
need some more benchmarking to see if this actually matters at
all. so far the results have been disappointing.

I'm fairly sure there was an open CLJS ticket for this kind of
stuff, but I can't find it. Either way its an experiment and
opt-in for now.