add experimental new shadow.lazy replacement for :target :esm
shadow.lazy relies on closure loader stuff which makes no sense
to use in a :esm setting. added the most minimal thing possible
so that it is possible to load a thing by name.

for this a given var needs to be tagged with ^:lazy-loadable

shadow.esm/load-by-name can load those names at runtime

so (esm/load-by-name 'some.ns/foo) will do a dynamic import()
and returns a promise that resolves to a zero arity function
that will return the value of that var. it doesn't just return
the value so that it works in hot-reload contexts.

moved the shadow-cljs UI to make use of this as a test